Created a function get_button_ordered_players()
What this function des is get all the players in game_controller in one array
ordered by the position of the button, with the button seat in the last position
of the list. This will make it much eaiser to control the flow of the hand because
the player at the beginning of the list is always first to act.