Setup linting and testing in travis CI and change all invalid code
A lot of small changes that did not pass python linting with
flake8. Backend now runs tests (there are 0 tests ready atm) and
linting. It also runs test with pytest in GameEngine.

Frontend now runs eslint on the project but is missing test runs,
to setup tests for frontend jest has to be setup and used.