Created create_superuser function and tested
Created a new function create_superuser, the difference between
super user and user is that is_superuser is set to True, but that
is a part of the PermissionMixin in djnago auth. Superusers will
be created via terminal and they will have power to change, delete
and manage other users.