Created management command for django wait_for_db
This command makes sure that the backend application will wait for
the database to be ready before running the django build.
Changed docker-compose file to use this command and also added
migrate command to the docker-compose file when the application is
started. The command is tested in test_commands.py