Created an endpoint for createing users
When user is created a url /api/user/create the user is added to
the database and a user object with only username and name is
returned to the user in the post response.

When a token login is posted at /api/user/token with username and
password an auto generated token is then returned to the user
which will then be used for access control on frontend and to
other endpoints.