Created a sign up form and tested it
Also experimented a lot with testing and changed a few things
around testing and css. Had to add some packages and update lint
and jest config.

The tests can be improved further, but there are a lot of problems
that come up when trying to make a good tests while using a UI
package like element-react.