Created the signup method in the Auth.js action file
The action uses axios to send a post request to /user/create route
in the backend and then dispatches SIGNUP_SUCCESS action or
SIGNUP_FAIL depending on the response.

The action should be fully tested with a mock-store and mock-axios
It tests  three cases, valid signup, empty fields sign up, and
multiple users sign up. The responses are simulation of the server
responses but not actual responses, because we don't make any real
requests to the backend in the tests.