Created actions for checkUser
The CheckUser function in user actions starts by checking if a the
token cookie exists in cookies, if it does it will set the
authorization header as the cookie value. Next it will send a
request to the backend that fetches the user if this is a valid
authorization token. Otherwise it will send back the error.

If the token does not exist it will set dispatch an unauthorized
user state, which is a guest (not logged in) user.