Created private and public routes and game route
Privite and public routes are made to distinguish between private
and public pages in the application. Logged in user should not be
able to see public routes and will be directed and the other way
around. Both can still see regular routes.

Created a game route, mostly to try this out but that route can
be used as a game menu for start.