Created action for initSocket, that creates the socket connection
The initSocket action takes in three parameters, the table because
each table should be on it's own socket url, the tokenString is
the authorisation token which is in the header of the socket
connection.

The last parameter is the store it self, the reason
the store is passed in as parameter is becuase to be able to
dispatch acitons to the store in .on functions of the socket we
need to have acess to the store at that point in time, which we
don't have unless we have a pointer to the store at that time.

The action is tested for a SOCKET_CREATED when initSocket is run
and for a single emit where an PLAYER_LIST action should be
dispatched.