Changed the socket setup in the project:
Removed socket.io-client which is not used anymore.

Added redux-dynamic-middlewares which makes it possible to add
middlewares on the fly, everytime a new socket connection is
created a new instance of redux-websocket(which uses native
WebSockets) middleware is created, this is done because each
connection has it own prefix (it's id), to be able to tell them a
part in the redux store and actions.

The socket reducers add the socket to the store when it is created
it keeps the connection url, and when it's open the open state is
set to true.

Tests updated for the new upgrades.