Setup socket connection with Token authorization on backend
When the frontend initializes a socket connection to the backend it
contains the token as parameter and the route as the table_id.

So when it connects to /ws/game/?Token=12345, the table_id is game and
the Token is 12345. This is then routed through a TokenAuthMiddleware
where it checks if the user exists. If the user exists the user is
added to the scope and the connection accepted, if not the connection
is closed.

There are also a few small modifaction in the backend, nothing big.
I also exposed the database port on the postgres docker container to
be able to connect to database locally.