Created table_gateway_consumer that communicates with game_controller
The Table Gateway Consumer is a django.channels worker that runs in it's
own docker container and speaks to other parts of the application via
the channel_layer (the redis layer with django channels).

Each table gateway consumer when the first player sits down at the first
table. Each table consumer can have from 1 to n many tables, but
keeping it in it's own container creates the possibility of having more
of them with a load balancer in between.

When a new table is added to the consumer it will create a new instance
of GameController that will control that table. Each controller runs on
it own separate thread at uses the standard Python 3 Queue to wait until
an event occurs. For example when a new player sits down at a table the
consumer will call a decorated version of game_controller.add_player().
The decorater will instead of calling add_player() add it to the Queue
and the thread will run it when it's free, and the consumer does not
have to wait at all after adding it to the function Queue. Each table
or controller can then broadcast results or state to all the player
consumers.

This way each Table Gateway Consumer can handle traffic for a lot more
of tables and just let the threads know what to do but not when to do it
or wait for it.

I also created the game_engine service that runs the consumer in
docker_compose, created and updated few tests and change testing file.