Displaying players in frontend
When a player sits goes to the url of the table, /game at this point.
It connects to the socket connection and the name of the player is
displayed in the frontend.

Created the reducer in the frontend for a message for the socket but it
will change a lot throughout the development process and still needs to
be tested.