Finished the ChipRack component, and tweaked settings for svg
The ChipRack component has a lot of settings variables in its state that
are used to calculate the positions of generated chips so they will
appear to lay in the racks of the table. Most of the initial state
variables should not be change.

When the component is mounted it will loop through a list of wanted chip
colors in each rack from state.componentList, it will generate randomly
between 1-19 chips of the same color in each rack. It uses an inline
style to position them correctly by the calculations.

Everything in the ChipRack is be fully tested.

Also did a few tweaks here and there regarding svgs and related subjects