Added the graphics for Player and stack
Each player is now put at a position at the table in a seat number from
1-9 the seat number will then use a switch case to create the correct
styleing to position the player graphic at the correct spot. The stack
amount and the player name will also show up in the graphics and a
visual stack graphic and an amount label will show up if the player has
a bet out. Most of this is done in percentages to be able to scale the
size of tables up and down as needed.

Everything is propType checked and unit tested.