Created the card class, no functionality implemented yet
The Card class is the class the displays the cards of the game. I added
the svg vectors into the render method of the class and now it's only
displaying the ace of spades at the bottom of the screen and needs to
be made dynamic. Added some of the styleing for the class and created
test to test the custom propType check for the card work correctly.

The class requires a card prop which is a string to tell the class which
card it should display. The card prop should be  formatted like "Ah" for
the Ace of hearts, all other formats are invalid ("AH" with a capital H
is also invalid.)