Finished Card class and created the Board class
The Card class renders with a Switch case that will render different
graphic for the card based on the card prop value. For example ace of
hearts ('Ah') will render the an visual Ace of hearts on the board.

The Board class is made for the community cards in the center of the
table. It takes in a list of cards of size 0-5 and renders the card
graphic to the table. The Board class also includes a stack that is used
as the pot, in the center of the table.

Everything should be fully unit tested.