Few changes to start using the Table model as a database model
The Game controller and consumer needed to associate each table with a
specific table from the Database. To this it was best to use the primary
key of the table. So now each table is on it own url table/TABLE_ID
where TABLE_ID is the tables primary key. This will need a a refactoring
at the frontend side later when we want to be able to support multiple
tables at a single page, but it does the job for now.

Also did a few cleanups with the Tests to start to support the new
implementation that uses the Table object from the Database and wrote a
few relevent tests.