Changes in player vs observer, and few communication updates
Created a difference between being an observer or player, this will be
necessary later own when players start buying in to seats.

Also created the protocol to keep the channel_id of both players and
observers in the game_controller so we will be able to send private
information to individuals.

Also updated the Table model and migrated:
Added: dealer_seat, sb_seat and bb_seat
This will all be part of the game_state and information that is good to
keep in the database if server goes down, so that table rotation can go
on in the correct order once the server comes up again.