Created the setup table function in game_controller
There is a lot of things happening in that function. It first makes sure
that there are at least two players ready to play. Then it determines
where the button/dealer_seat should be positioned and after that sb, bb.
It also updates the game_state after it's calculations and the table
model in the database.

Also added a schema validation for the render method, that can also work
as a general validation of the state of the table.

Deleted the Table class which was replaced with the Table model.

Created tests and few minor fixes.