Finished player action and the control functions for the action queue
Created the function get_possible_actions in Player that takes in what
is currently the biggest bet on the table, how much the last_legal_raise
was and if all other players are all_in, the function than uses this
information to determine what actions the current player can do. It's an
important function in the game flow and therefor very well tested in
test_player.

In the hand class I also created a action_queue that will control what
players do and in what oreder. There are two functions that control the
action queue fill_action_queue and update_action_queue:
* fill_action_queue: is used for example when a flop has been dealt and
all active players should be in the queue.
* update_action_queue: is used when a bet or raise happens and players
need the chance to action.