Finished the player_action method in hand and created update_table
The player_action method updates the relevant variables like
current_max_bet, the pot and last legal raise. Wrote thorugh tests for
it as well.

The update_table method is the one that extracts the next player to do
from the action queue, or let the controller know that the hands need
to be ended. It also takes the hand to next stage when needed.