Updated the docker image for the backend to Python3.8 'Buster'
Buster is a debian images that comes with all needed packages
preinstalled so some install lines could be removed from the dockerfile.
But the main reason for the change is that alpine has been proved to
have some performance issues, especially with threading which is why it
is the best choice for this application.